chore: leftover gambit references on the backend
