+++ framework/core/src/Discussion/Search/Filter/UnreadFilter.php
@@ -7,7 +7,7 @@
  * LICENSE file that was distributed with this source code.
  */
 
-namespace Flarum\Discussion\Filter;
+namespace Flarum\Discussion\Search\Filter;
 
 use Flarum\Discussion\DiscussionRepository;
 use Flarum\Search\FilterInterface;
