+++ framework/core/src/Search/SearchServiceProvider.php
@@ -11,17 +11,17 @@ namespace Flarum\Search;
 
 use Flarum\Discussion\Filter as DiscussionFilter;
 use Flarum\Discussion\Search\DiscussionSearcher;
-use Flarum\Discussion\Search\Gambit\FulltextGambit as DiscussionFulltextFilter;
+use Flarum\Discussion\Search\FulltextFilter as DiscussionFulltextFilter;
 use Flarum\Foundation\AbstractServiceProvider;
 use Flarum\Foundation\ContainerUtil;
 use Flarum\Group\Filter as GroupFilter;
 use Flarum\Group\Filter\GroupSearcher;
-use Flarum\Http\Filter\AccessTokenSearcher;
 use Flarum\Http\Filter as HttpFilter;
+use Flarum\Http\Filter\AccessTokenSearcher;
 use Flarum\Post\Filter as PostFilter;
 use Flarum\Post\Filter\PostSearcher;
 use Flarum\User\Filter as UserFilter;
-use Flarum\User\Search\Gambit\FulltextGambit as UserFulltextFilter;
+use Flarum\User\Search\FulltextFilter as UserFulltextFilter;
 use Flarum\User\Search\UserSearcher;
 use Illuminate\Contracts\Container\Container;
 use Illuminate\Support\Arr;
@@ -43,14 +43,14 @@ class SearchServiceProvider extends AbstractServiceProvider
                     HttpFilter\UserFilter::class,
                 ],
                 DiscussionSearcher::class => [
-                    DiscussionFilter\AuthorFilter::class,
-                    DiscussionFilter\CreatedFilter::class,
-                    DiscussionFilter\HiddenFilter::class,
-                    DiscussionFilter\UnreadFilter::class,
+                    \Flarum\Discussion\Search\Filter\AuthorFilter::class,
+                    \Flarum\Discussion\Search\Filter\CreatedFilter::class,
+                    \Flarum\Discussion\Search\Filter\HiddenFilter::class,
+                    \Flarum\Discussion\Search\Filter\UnreadFilter::class,
                 ],
                 UserSearcher::class => [
-                    UserFilter\EmailFilter::class,
-                    UserFilter\GroupFilter::class,
+                    \Flarum\User\Search\Filter\EmailFilter::class,
+                    \Flarum\User\Search\Filter\GroupFilter::class,
                 ],
                 GroupSearcher::class => [
                     GroupFilter\HiddenFilter::class,
