+++ framework/core/src/User/Search/Filter/EmailFilter.php
@@ -7,7 +7,7 @@
  * LICENSE file that was distributed with this source code.
  */
 
-namespace Flarum\User\Filter;
+namespace Flarum\User\Search\Filter;
 
 use Flarum\Search\FilterInterface;
 use Flarum\Search\SearchState;
