+++ framework/core/src/User/Search/Filter/GroupFilter.php
@@ -7,7 +7,7 @@
  * LICENSE file that was distributed with this source code.
  */
 
-namespace Flarum\User\Filter;
+namespace Flarum\User\Search\Filter;
 
 use Flarum\Group\Group;
 use Flarum\Search\FilterInterface;
